Update patch set 4
Patch Set 4:

> > Patch Set 4: Code-Review-1
 > >
 > > (5 comments)
 > >
 > > We really should find a way to not have to duplicate 90% of a
 > file in another suite. A starter might be to just move everything
 > to some keywords that are shred between the suites. That at least
 > keeps the code in one place, though you still need to dup the calls
 > of the keywords in the tests.
 > 
 > I was thinking the same thing, but things are pretty compact as is.
 > I mean
 > we could probably wrap the first 3-4 test cases in to a single
 > keyword. the
 > ones that create networks, subnets, etc. But after that, in this
 > suite it
 > diverges into blocking ports before creating VMs, etc. I don't like
 > copy/paste
 > suites either, but at some point we can't collapse shared work too
 > much
 > lower. My other idea was to just add the relevant work from this
 > suite
 > (isolation of ODL nodes via blocking clustering ports), and it
 > might work,
 > but it also starts to make a really heavy single suite. what do you
 > think?
My first thought wasn't too collapse multiple tests, but to just move most of the code to keywords that could be called by the tests. Then the bulk of the code is in the keywords and the tests are just steps. Not great but seemed better than duping most of the code.

If it is just adding some blocking tests then we could also easily just exclude the tests in the main run. That might actually be the cleanest and easiest. Seems like that could be easy to do.

Patch-set: 4