Update patch set 8
Patch Set 8:

The failure in the previous run is due to genius 3 node CSIT issues, where ovsdb plugin is not able to handle fast reconnects and connects. Vishal says, genius 3 node CSIT has to be rewritten to use HAPROXY and all, after that only we will be able to check if this is a genuine failure. Anyways, this failure is not related to the current patch.

Patch-set: 8