Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > (1 comment)
>  > 
>  > we need to double check this is working in the sandbox, btw.
> 
> Aswin noted that the pnf flows were eventually installed after 12s, but the ping is just 3s. so I increased it to 30 count to give us 30s.

the sandbox failed even with a 30 count. either way, I think we gotta keep this
as a failure and blocker bug to get figured out. If we find a delay gives us a
passing test case and workaround, then we'll never get the bug fixed. it's a new
problem right?

Patch-set: 2