Allow offsetting index for raft queries
Our downstream CI uses tripleO for deployment of OS. This can cause
differences in deployment. In this case it's related to akka.conf
which indexes member roles from 0, instead of 1. This results in
some queries failing with 404.

This patch allows us to add an offset for indexes so that we query
the appropriate member-${index}

Change-Id: I024eb521436530806c15fd4109ddf3644a2191f4
Signed-off-by: Tomas Jamrisko <tjamrisk@redhat.com>