Update patch set 2
Patch Set 2: Code-Review+2

> Patch Set 2:
> 
> > Thanh, is there anything needed besides this to keep the docs
>  > from being updated automatically?
> 
> Not sure I understand. You don't want the docs to be updated automatically?
> 
> The current code here will automatically update docs in RTD with every rtd merge job run. (We will need to tie that in releng/builder but this needs to go in first).

what I'm asking is how it will all work automatically now that we
are moving the scripts to this repo. It was working automatically
before, so I figured we at least would have to point the jobs/tooling
to this repo now. I don't really know the details that well, but 
since you say we still need to tie it in to releng/builder, I assume
we'll need a patch for that.

can you let me know what to do for that?

merging this now.

Patch-set: 2
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+2