Use Variables.robot in two main libraries
We need to keep moving on this conversion. This patch
is to fix a bug that was very irritating to track down.

this line in Variables.robot is pretty simple:
  @{AUTH}           ${ODL_RESTCONF_USER}    ${ODL_RESTCONF_PASSWORD}

But, changing the two ODL_ variables on the command line
so that we can have AUTH be something different for
other installations was not working. AUTH was still taking
admin:admin as user and password.

That's because we were still sourcing Variables.py in
these two libs (Utils and TemplatedRequests) where this
line exists:

 AUTH = [u'admin', u'admin']

pybot brings in those .py variables files last and they will
take final precendence, even over the command line overrides.

very annoying, and also very hard to figure out unless
you are thinking about it.

Change-Id: I0b91be872ab5dd7a598a7b13c8b6bb45e6b986f0
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>