Update patch set 73
Patch Set 73:

(4 comments)

I started to comment for each problem for the -verify, but it'll be easier
if you do this:

'python -m robot.tidy <filename>' to get things the way the job is expecting it. that will put the corrected file to stdout and you can just overwrite the existing file. I normally put it to a tmp file so I can run a 'diff' and see
what stupid things it's complaining about first.

Patch-set: 73
Label: Verified=0