Update patch set 10
Patch Set 10:

(2 comments)

> Problem Recreated for fixing
 > 
 > Sandbox: https://logs.opendaylight.org/sandbox/vex-yul-odl-jenkins-2/netvirt-csit-1node-openstack-pike-upstream-stateful-snat-conntrack-nitrogen/3/robot-plugin/log_03_external_network.html.gz

Venkat,

some cosmetic things to add (sorry I missed them before), but the
bigger question is if this is expected to pass in Oxy and/or
Fluorine. Is it just failing in Nitrogen?

We are trying to keep our main jobs passing 100% to keep the
confidence of devs using them as gates. If it's going to fail
right now, we can hold off on merging it, or we can find a way
to put this suite/test-case in the tdd-wip job(s)

Patch-set: 10