Update patch set 1
Patch Set 1: Code-Review+1

> Our downstream jobs are failing because the undercloud can't
 > connect to karaf, as 8101 is blocked by default. I could allow it
 > in infrared-robot, if you think that's a better place

no, I mean to add this info as part of the commit message. Just
write that in a typical tripleO deployment, the controller node
does not have the karaf port (8101) accepting connections, so we
need this in order to more easily run CSIT on tripleO deployments.

doing some sandbox double checks, just to be safe.

here:
  https://jenkins.opendaylight.org/sandbox/job/jamo-netvirt-csit-1node-openstack-pike-upstream-stateful-snat-conntrack-oxygen/

and here:
  https://jenkins.opendaylight.org/sandbox/job/jamo-netvirt-csit-3node-openstack-pike-upstream-stateful-snat-conntrack-oxygen/

Patch-set: 1