Update patch set 2
Patch Set 2:

the genius-upstream.txt content was replaced with a shell script rather than a testplan. Notice the other testplan.txt files are pointing to the file and they eventually expect a list of tests to run. Thet genius-upstream.txt file is filled with the shell script though. 

The script in the genius-upstream.txt should be in a different .sh file. Once of those .txt testplan files would have the path to that script in in, then all the other .txt files would softlink to that .txt testplan.

At this point, I would just scrap the whole link stuff to get the tests running again. It's ok to repeat the line in the testplan. Then we can add links in a later patch.

Patch-set: 2