Update patch set 31
Patch Set 31:

but, it is not even accepting that genius-verify.txt in verify jobs, as verify jobs fail to find any test suite in verify tesplan when we do like that.

So, for the time being, shall i just copy the new suite in genius-verify.text and upload? Am  not finding any other way

Patch-set: 31
Label: Verified=0