Update patch set 28
Patch Set 28: Code-Review-2

(6 comments)

-2 because of the commented testplan.

can you also run the full job with all the suites with your
updated patch. just a few nits about putting a space and "="
in a few variable assignments. I also had a more technical
question that I was curious about.

it's a good suite. we should be able to get this in soon.

Patch-set: 28
Label: Code-Review=-2