Update patch set 17
Patch Set 17:

Hi Jamo,
The tox verification is failing with the below error:
The commit message body (all lines but the first line) contains lines that are too long. The should be 72 characters or less with the possible exception of unbreakable things like URLs. Please update the commit message as necessary to pass this check.

It is failing now because I have added one more line in commit message,i.e,
copy-logs: netvirt-csit-1node-openstack-queens-madhuupstream-stateful-oxygen/2
as per your suggestion to save the logs in server.

Again, I shared the pass log for oxygen with pike, in which nano flavor is not available. So, need to create nano flavor in this.
But in oxygen with queens, nano flavor already exists.Due to which script is failing.

Please suggest how to proceed.

Thanks,
Madhusmita

Patch-set: 17