Update patch set 17
Patch Set 17:

> Hi Jamo,
 > The tox verification is failing with the below error:
 > The commit message body (all lines but the first line) contains
 > lines that are too long. The should be 72 characters or less with
 > the possible exception of unbreakable things like URLs. Please
 > update the commit message as necessary to pass this check.
 > 
 > It is failing now because I have added one more line in commit
 > message,i.e,
 > copy-logs: netvirt-csit-1node-openstack-queens-madhuupstream-stateful-oxygen/2
 > as per your suggestion to save the logs in server.
 

copy-logs is a comment to be used in a gerrit comment to trigger a job
to copy the logs. it's not for the git commit message.

 > Again, I shared the pass log for oxygen with pike, in which nano
 > flavor is not available. So, need to create nano flavor in this.
 > But in oxygen with queens, nano flavor already exists.Due to which
 > script is failing.
 
I *think* that when we ask for tempest to be installed with devstack
we get the m1.tiny image for free in pike and queens. I just ran
two jobs in the sandbox (pike and queens) and had it do an openstack
flavor list before starting robot. I see m1.tiny there in both:

https://jenkins.opendaylight.org/sandbox/job/jamo-netvirt-csit-1node-openstack-pike-upstream-stateful-oxygen/1/consoleFull

https://jenkins.opendaylight.org/sandbox/job/jamo-netvirt-csit-1node-openstack-queens-upstream-stateful-oxygen/5/consoleFull

do you have a job showing this not working in pike?

if you look at the first suite that runs in these jobs (the l2.robot suite)
you can see that it's creating it's instances with that m1.nano
flavor in all the jobs.

anyway, show me the job with trouble and we can figure out what
to do.

JamO

 > Please suggest how to proceed.
 > 
 > Thanks,
 > Madhusmita

Patch-set: 17