Update patch set 2
Patch Set 2:

I believe at the time I set the value to 50 (in [0]), the value has been reached, at least in longevity tests.
Now I see the longevity tests (such as [1]) are not being run, so we do not know if there was a regression, or the short test is still suffering from being run while ODL is busy staring the -all- components.

My slight preference is to keep the test failing, as a reminder we do not have a reliable performance-readiness check in place.
But I have added Robert as a reviewer, to decide which signal the developers are interested in.

[0] https://git.opendaylight.org/gerrit/63044
[1] https://jenkins.opendaylight.org/releng/view/controller/job/controller-csit-3node-cs-chasing-leader-longevity-only-fluorine/

Patch-set: 2