Update patch set 80
Patch Set 80: Code-Review-1

(3 comments)

I think this is really close now. just two small nits I just posted
a comment on.

Please run the full jobs (not just this suite) in the sandbox. Please
do it for oxygen as well as fluorine. Also make sure to run the
snat-conntrack version of the job (in addition). Once the jobs are
done and you are satisfied, please do the copy-logs: trick so they
are saved for good. I finally found time to get some reviews in 
today, but the previous log posted from the sandbox has been purged.
the sandbox is totally purged (jobs, logs, etc) every weekend.

thanks :)

Patch-set: 80
Label: Code-Review=-1