Update patch set 4
Patch Set 4: Code-Review-1

Is this patch the same as https://git.opendaylight.org/gerrit/#/c/72227/ or is this totally different? Some of the changed files are the same. If it is the same file it would be better to abandon this patch here and then update the original at 72227 so we can keep history.

Patch-set: 4
Label: Code-Review=-1