Update patch set 72
Patch Set 72:

this is ready then? can you give a link to the good sandbox
logs? Looks like the last log given was direct to the sandbox
which is purged every weekend, so I cannot see it. hoping
you used the copy-logs trick.

also, a question about the configplan which is now going to change
the default of bundle based resync from false to true. Unless
I'm missing something, that is going to affect all of our netvirt
CSIT jobs going forward. Is this what we want?

Patch-set: 72