Update patch set 11
Patch Set 11:

> This has started running with the new flag enabled, and there are
 > test failures. Should we wait for all the issues to be fixed before
 > this job can get in? Or can we get this in, and test our pending
 > patches against this and stabilize? copied the logs for reference

imho, since this is just a *new* job and not going to affect the normal
job we are looking at for stability and using for gates, we can get it
in and debug from there. Seems like there are legit bugs (from irc
conversation with Faseela) and people will need to work to stabilize
things.

Patch-set: 11
Label: Verified=0