Update patch set 1
Patch Set 1:

> in this patch, i have updated this keyword. So, can we abandon
 > this?

well, what's wrong with this patch?
Is it addressing anything else?

If it's adding something you are also adding in another patch, that
is no problem. We could merge this and you can rebase and we'll be
fine.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0