Update patch set 12
Patch Set 12: Code-Review+1

robot code flies with me. sandbox jobs were good
for what this added. I was surprised to see that
exception failure for the karaf ssh stuff. That shouldn't
be showing up as we disable that log.

any idea Sam?

unless we want to double check with pike and/or oxygen, I'm
fine to merge.

looks like a good bit of extra coverage coming in.

Patch-set: 12
Label: Code-Review=+1