Update patch set 1
Patch Set 1:

> > https://git.opendaylight.org/gerrit/#/c/71990/
 > 
 > I think we can get a more efficient git/gerrit process here.
 > this patch is super small and only requires a small change
 > in a single keyword. correct me if I'm wrong.
 > 
 > the patch above c/71990 is large and probably will take
 > some time.
 > 
 > why not merge https://git.opendaylight.org/gerrit/#/c/72861/ or
 > just add the keyword changes to this small patch.
 > 
 > you can then rebase c/71990 after that and everyone will
 > be happy.

you are correct jamo. But if you take a look we are using NUM_TOOLS_SYSTEM which is 3 but we use only two TOOLS_SYSTEM_IP so which will fail. So the patch above c/71990 is to enhance the whole genius to 3 switches in the scene where we use 3 TOOLS_SYSTEM_IP's. So I made this patch part of that c/71990 patch. So that we don't break any current GENIUS-CSIT. :)

Patch-set: 1