Update patch set 14
Patch Set 14:

> I have verified that it doesn't break any existing jobs, though the
 > new job fails on all patches till one netvirt patch goes in.. I am
 > okay to wait till that patch goes in, to run this again and see if
 > it helps..I was just worried about the overhead of keeping on
 > rebasing these patches till all the fixes get in :P But it is
 > okay..

This does produce a number of exceptions in the job. They are tunnel related so likely related to these patches. Can you confirm this is because the genius side is not there?

https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/builder-copy-sandbox-logs/107/Karthikanetvirt-csit-1node-openstack-queens-upstream-stateful-itm-direct-tunnels-fluorine/9/odl_1/odl1_exceptions.txt.gz

Either way, you should run this job with the enable flag as false so we can verify the existing jobs are fine. For that I started https://jenkins.opendaylight.org/sandbox/job/Karthikanetvirt-csit-1node-openstack-queens-upstream-stateful-itm-direct-tunnels-fluorine/10

And then run this patch with the build from https://git.opendaylight.org/gerrit/#/c/72867/ to see if the new stuff works. just started that build so need to wait till it is ready.

Patch-set: 14
Label: Verified=0