Update patch set 1
Patch Set 1: Code-Review-2

(6 comments)

I bet if you only include the first portion of this suite
up to when it brings down and back up the first controller,
we can review and merge this a lot faster. That will give
you a base to build on for adding the test cases to bring
down the 2nd and 3rd controllers, as well as multiple controllers.

Patch-set: 1
Reviewer: Gerrit User 2264 <2264@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-2