Update patch set 3
Patch Set 3:

> Jamo, currently genius has a common 1 node and 3 node test plan..
 > with this ha test suite, we have to split that up. What is the
 > identification for 3 node test plan? How should we name the file?

you can override the testplan name in the job definition if you
want to create a special one to use. but, you are using the
inttest-csit-1node and inttest-csit-3node in the same "project"
of the genius-csit-genius.yaml file.

probably simplest to make two projects in that file. one for
each of those. then just specify the exact testplan file in
each of those.

btw, by default CSIT jobs first look for a testplan file named:
{project}-{functionality}.txt if nothing is specified. just an
FYI

Patch-set: 3