Update patch set 1
Patch Set 1:

> (1 comment)

This change is part of the exercise to check if deleting tun-ports is really needed now.

Tun-ports were being deleted originally because of interface creation dependency in Genius. Testing updates in TripleO has revealed (as part of bug discovery) that deletion is not needed as update is successful even without deleting.

There was some work going on in Genius to remove this dependency and looks like that work has landed in, which will mean that we will really be removing this part of code from update process completely.

Patch-set: 1