Update patch set 17
Patch Set 17:

> Hi Jamo,
 > 
 > All review comments are implemented.
 > Can you please review this current patch.
 > 
 > Log :
 > https://logs.opendaylight.org/sandbox/vex-yul-odl-jenkins-2/netvirt-csit-1node-openstack-queens-sanjoupstream-stateful-fluorine/9/
 > 
 > In this log, one merged testcase is failing which is because of the
 > build issue. Other testcases are passing.
 > Please look into it and let me know your comments.
 > 
 > Thanks,
 > Madhusmita


so we need to understand the failure. is there a bug? is someone
working to fix it? We don't want to add new failures to our
stable job at this point. We use it to gate dev patches. We can
add this suite to the tdd-wip job if needed.

Either way, when you run the sandbox jobs at the point when you
think it's close to being done, please run the full job and
not just these suites. We need to make sure we see that everything
is working well together.

Patch-set: 17