Update patch set 34
Patch Set 34:

(8 comments)

Hi Jamo,
I have implemented all your review comments.

But for one comment "Not to use “OpenStackOperations.Create Nano Flavor”. ", I have raised a concern and shared the log with you.

And as this is a tdd-wip job, so we don't need to check the main netvirt job combos,
 Please find the pass log :
 1.  https://logs.opendaylight.org/sandbox/vex-yul-odl-jenkins-2/netvirt-csit-hwvtep-1node-openstack-queens-anujupstream-stateful-fluorine/6
 2.  https://logs.opendaylight.org/sandbox/vex-yul-odl-jenkins-2/netvirt-csit-hwvtep-1node-openstack-queens-anujupstream-stateful-fluorine/7

Thanks,
Madhusmita

Patch-set: 34