Update patch set 30
Patch Set 30:

please rebase this patch so that it's a child of c/72445

that will make the change you have in OVSDB.robot and
OpenStackOperations.robot go away.

another way to have worked through this, is to first make a patch
with the two lib changes, then base your vni l2 and vni l3 patches
on top of that.

Patch-set: 30