Update patch set 18
Patch Set 18:

> @jamo: are some of the changes things we want to push to other
 > clustering suites?

no, there isn't really anything here that will apply to other
suites, with one exception (BTW, learning a lot while doing this
work):

- when we lose a node (leader only, I think) and transactions
were in flight, we could end up (ask) timing out on that transaction
and if it was made from a REST call, that time out will bubble
all the way back to the client as a 500 error. This is expected
for ask-based. So, if we find those cases in other suites we can
make sure to add a retry and be OK with a few 500s during leadership
changes.

NOTE: this is not expected for tell-based, as tell-based itself
will do it's internal retries so the REST client can eventually expect
the good 2xx response.

OH, there is still a problem that we can lose /restconf to 401 for
quite a while (~1m) during a leadership change, but that will be
fixed when Robert's patches go in:

https://git.opendaylight.org/gerrit/#/c/73475/

Patch-set: 18