Update patch set 68
Patch Set 68:

> (1 comment)
 > 
 > > Hi vishal,
 > > It has done in generic way where we have n switches. But in
 > jenkins
 > > we have only 3 switches so named as 3 switches but I have made it
 > > such a way it works for n. Please review and post your comments
 > :)
 > 
 > I noticed robot scripts are generic, but json to create teps is not

Vishal, Can we review the current changes, and take up the REST call generalization in subsequent patch. Sathwik has tried to make changes as generic as possible, the JSON was also part of initial agenda, but asked him to speed up and atleast get the current thing for 3 node done, and then take up the JSON genralization in a subsequent patch.

Patch-set: 68