Update patch set 2
Patch Set 2:

Hi Luis,
I am validating this new keywords for the suite :
"Add Subnet And Multicast" [https://git.opendaylight.org/gerrit/#/c/72152/]

This is a child patch created for the parent patch 72152.
As per Sam, because of the new keywords adding in same library file for different suites,the reviews are becoming more difficult and takes longer. 

So, to avoid this, we are creating differnt parent and child patches.
 

Thanks,
Madhusmita

Patch-set: 2
Label: Verified=0