Update patch set 3
Patch Set 3:

this is dependent on c/73917/3 which I think was not intended.

I think we only need to make reviews dependent on each other if
they are building up on some common new changes. So, like the
test cases that would use the BgpOperations keywords added here would
be dependent on this patch, but there isn't any reason for this
review to be dependent on the new keywords you are adding to
VpnOperations. right?

I'm rebasing this to get out from under it.

Patch-set: 3
Label: Code-Review=0