Update patch set 1
Patch Set 1:

JamO, yes agree. I really want to see if the results are any different for controller csit with this change in the tests that I have been running.

This patch doesn't change all of the instances of "kill -9" in controller csit, for example, in the cars csit test that you have been running, is not included in this WIP patch.

Patch-set: 1