Update patch set 9
Patch Set 9:

Hi @Luis. Thanks for the review! Yeah. I will not only add the documentation for the format of the config files, I'll also add some documentation for the functions I have used in the scripts. I have removed the trailing spaces. And since there is nothing more to be done for now in this, I think this can be merged. I'll add the docs in the next patch.

Patch-set: 9