Update patch set 14
Patch Set 14:

Also, if you will notice, in the dashboard view, Config DS has no lines in between the points even though in the individual visualization view it is present as it should be for both time and rate. The problems  sometimes goes away when refreshing the dashboard. I believe this might be an issue with the Kibana.

Patch-set: 14
Label: Verified=0