Update patch set 2
Patch Set 2:

> @jamo: this change breaks the csit. Looks like name is hard coded
 > at some places.
 > 
 > https://logs.opendaylight.org/sandbox/vex-yul-odl-jenkins-2/thapar-controller-csit-3node-rest-clust-cars-perf-ask-only-fluorine/1/robot-plugin/log.html.gz#s1-s2-k1-k11-k1-k1-k6-k1

Vishal, the problem here is the rest call to get leadership for the
car shard returned some trouble. it got this:

https://logs.opendaylight.org/sandbox/vex-yul-odl-jenkins-2/thapar-controller-csit-3node-rest-clust-cars-perf-ask-only-fluorine/1/robot-plugin/log.html.gz#s1-s2-k1-k11-k1-k1-k6-k1

instead of something like this:

https://logs.opendaylight.org/releng/vex-yul-odl-jenkins-1/controller-csit-3node-rest-clust-cars-perf-only-oxygen/341/robot-plugin/log.html.gz#s1-s2-k1-k11-k1-k1-k6-k1-k1-k1-k6-k1-k4-k1


This is weird. the featuresBoot is the same here, and the first suite ran ok.

is it possible we just hit a bug? that was the first build in that sandbox
job. did you run others that failed the same way?

Patch-set: 2