Update patch set 7
Patch Set 7:

given the history of this work before and the failures that
poppped up in our releng csit jobs, can we please get multiple
sandbox runs? make sure to cover all combos. in other words:

oxygen and fluorine
pike and queens
snat-conntrack and not

Thanks for all the hard work and debugging.

Patch-set: 7