Update patch set 2
Patch Set 2:

> @Jamo: If you look at test cases, you will see that ask and tell
 > based tests are not same. There are some tests that are ask only
 > and some that are tell only. Josh's point is as part of this
 > activity we should also consider shoring up both to make suites
 > identical except for ask vs tell. You can take a look at
 > https://git.opendaylight.org/gerrit/74692
 > 
 > We can consider doing this in a separate patch later too as I think
 > not all tests maybe applicable to both.

ok, but this patch is the right place to start right? just start with
the suites that we know were intended for each protocol, and make
two jobs. Then we can figure out and add the other suites between
jobs as applicable.

even so, what I've done here resulted in too many failures I think.
so, something is wrong. note that I'm not doing the restart between
suites, so maybe that's a requirement? anyone know?

Patch-set: 2