Change OVS restart to use restart cmd
Using the start and stop commands results in 2 commands, where if the
robot connection to the node is via the same IP used on an OVS bridge,
will sever the connection on the first command. Using the 'restart'
command instead preserves the connection and restarts OVS in a single
command.

Change-Id: I05b234ed95e5456fbb53fbb4fea695802883cf48
Signed-off-by: Tim Rozet <trozet@redhat.com>