Update patch set 34
Patch Set 34:

> (1 comment)

OK, maybe that is the confusion. The current netvirt csit upstream/snat-conntrack jobs all have l3vpn tests in them but none have a tools system vm. Possibly there are releng/builder patches where tools_system vms are being added to the netvirt-csit-multi-openstack.yaml files. Is the plan to create new job variants that will need the dcgw-setup.sh? If so, then we should add a job parameter flag to include the dcgw-setup for those jobs. That can then be used for the dcgw-setup.

Patch-set: 34