Update patch set 18
Patch Set 18: Code-Review-1

(1 comment)

Very minor comment to address, otherwise I think this looks good.

Does this patch depend on the dcgw-setup.sh stuff? If so can you add the Depends-On: <Changed-ID> to the commit message.

Can this patch go in independent of the dcgw-setup.sh? If so, then no need to add Depends-On, but still update the commit message to say that Change-ID is something expected for the test suite to truly be functional.

Patch-set: 18
Label: Code-Review=-1