Remove basic test suites
These suites were never meant to be kept and were templates
for projects to get started. From there the projects were
to add real tests. As they are now, they do not test anything
related to the project and simply use resources.

Long-term new projects should use the templates in the sandbox
to gain experience with robot and csit and then update to a real
test before merged in production.

Change-Id: I7e097f30d118428cf468e4e61d7e9b33e25c3d86
Signed-off-by: Sam Hague <shague@redhat.com>