Fix for OPNFLWPLUG-1022 requires Topology Manager
module to register for singleton clustering service.
That introduces application entity in the EOS output.
In the current cluster tests, after disconnecting all
the OpenFlow devices, test checks that no openflow
 device should have owner, by grepping the "member"
string, and expect that it should be 0. After
introducing topology manager entity, EOS output
will always have one entity with all the 3 controllers
as a candidates, so greppign the "member" string
will always return 3. This patch take care of this
change."

Change-Id: Id2d3c0905ec3cd02c2098c1c1aee61c081dfe629
Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>