Update patch set 4
Patch Set 4:

> > Vishal, can you pick these back up again? maybe get some jobs in
 > > the sandbox and let's take a look. If we can get these working
 > and
 > > merged it will hopefully clarify some more problems.
 > >
 > > btw, partner patch is https://git.opendaylight.org/gerrit/#/c/75014
 > 
 > Done:
 > https://jenkins.opendaylight.org/sandbox/user/thapar/my-views/view/MyView/job/thapar-controller-csit-3node-clustering-tell-all-oxygen/
 > https://jenkins.opendaylight.org/sandbox/user/thapar/my-views/view/MyView/job/thapar-controller-csit-3node-clustering-ask-all-oxygen/
 > 
 > Do we want to switch to flourine now?

no, maybe you missed it in the meeting, but we are keeping focus
on Oxygen until SR4 is released and then we will move to Fluorine.

I copied your job #1 sandbox logs and sent an email asking for
some help figuring out which suites need to stay in which job.

https://lists.opendaylight.org/pipermail/controller-dev/2018-August/014605.html

Patch-set: 4