Update patch set 2
Patch Set 2:

> patch is cherry-picked on all the branch - stable/fluorine,
 > stable/oxygen.
 > 
 > BTW, it's very inconvenient that we have only one integration/test
 > master branch that we use for all the project branches. Functional
 > behavior can change from one release to other release and
 > individual test might need a change or sometimes the entire suite
 > might need a change, and i think using tags to make sure of that
 > probably is a very complex approach. I was thinking i will
 > cherry-pick this csit patch to stable/fluorine and stable/oxygen of
 > integration/test(just like any other normal project), but because
 > of the current integration/test single branch, i need to make sure
 > that OFP project merge all the three patches together, which means
 > i might have to wait till Fluorine branch is unlock. Because if i
 > will only merge to OFP master, than master branch gate job will
 > fail, and to make it green i will have to merge the csit patch, but
 > if i will do that, than my Fluorine branch gate test will fail :-/,
 > not a good motivation to contribute CSIT test to integration/test
 > ;-P

you are right, of course.

this is just such a rare case for int/test that it was not, and still
not worth the overhead of maintaining branches. Probably once a year
we hear this concern, but if it becomes more common we can surely
switch to tracking branches like the other projects.

Luis, I think we need to step in and help Anil here. Can you do that?
Last thing we want is finally getting a legit java dev pushing robot
code and scaring him away :)

Patch-set: 2