Update patch set 8
Patch Set 8:

The "Chasing The Leader" test fails sporadically with :
"Acceptance rate 50 not reached, actual rate is 45.4333333333"

So it's expecting a specific rate for leader changes per sec. I don't know who picked 50 as the magic number or why but I think it should just report the rate and not fail  if some  expected rate isn't reached - that will never be reliable in the jenkins env unless the threshold is really low. 

Just noting that here for a future patch....

Patch-set: 8