Update patch set 8
Patch Set 8:

> The "Chasing The Leader" test fails sporadically with :
 > "Acceptance rate 50 not reached, actual rate is 45.4333333333"
 > 
 > So it's expecting a specific rate for leader changes per sec. I
 > don't know who picked 50 as the magic number or why but I think it
 > should just report the rate and not fail  if some  expected rate
 > isn't reached - that will never be reliable in the jenkins env
 > unless the threshold is really low.
 > 
 > Just noting that here for a future patch....

I think Vratko picked 50 as a safe number, and we've seen this over
100 before. Having this test and having it fail already helped us
once. Originally I worked on a patch to lower the rate to something
more safe, but that caught the attention of Robert and Vratko and
we realized we missed an actual cherry pick to remove some
sleeps in the controller project.

this gerrit has some conversations about this:
https://git.opendaylight.org/gerrit/#/c/72136/

We can lower it more, but I'm wondering if 47 is not so far beneath
100+ that we have seen before that we don't actually have an issue
to figure out.

But, if we are just going to report a rate and not pass/fail on it,
and we don't have anyone ever paying attention to it, it's pointless.

Tom, do you understand what the test is doing and what the rate
means? Can we make an educated guess on what the value should be?

Patch-set: 8