Ignore ssh errors on closed connections
Commit I7bb0f566eef9c4e5526272a2af567d0180db8e09 brought in a
change to logout from ssh connections, but in some cases the
connection is gone, such as during stop and start of ODL nodes
in clustering tests. This then fails the keyword and exits
so the later part of the keyword doesn't execute. And that is
the important part which opens a new connection. This ends up
leaving the old connection which is closed and unusable.

Change-Id: I3512686e7df117c453789608bc5c3b2121e7c7b8
Signed-off-by: Sam Hague <shague@redhat.com>