Update patch set 9
Patch Set 9:

> I know what you are saying about the unpredictability of our infra,
 > but we should still be able to find some safe value for the most
 > part right?
 > 
 > This graphic might help make my point:
 > 
 > https://docs.google.com/spreadsheets/d/17ztFMkcRbSwPn3St-NhZcE7CWVCi4o1tAce720SbRkA/edit#gid=0
 > 
 > These are the last ~150 jobs worth of rates (two per job)
 > It should be pretty obvious when we finally realized we were
 > failing
 > and got a fix in place.
 > 
 > btw, it's not for certain, but looking at the graph in the link
 > above,
 > it's almost like we have a downward trend. Not asking for anyone to
 > care,
 > just pointing it out.

But remember the fix was actually in the test code itself - there was an artificial sleep FWIC. So that was only relevant b/c CSIT was thresholding on the rate.

I think it's curious why singleton service change rate was chosen for thresholding anyway. Why don't we similarly threshold shard leader change rate? Or transaction rate?

Anyway, the rate is heavily dependent on the network communication performance between the nodes which is why it's so variable and why it's difficult to threshold it. Looking at the data, 20 seems like a pretty safe low value but even then there was one anomalous run of 7.9. So if we pick something like 20, we have to be OK with say possibly 1 out of 50 runs failing and ignore it - if it started to become a trend then we know there's an issue.

Patch-set: 9